% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClasses.R
\name{addRcppMethod}
\alias{addRcppMethod}
\title{Add Rcpp Method in \code{\link{testClass}}}
\usage{
addRcppMethod(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite methods of the name name?}
}
\description{
Method for testing
}
