% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testClasses.R
\docType{class}
\name{testClass}
\alias{testClass}
\title{Test Class for Testing Rcpp Methods in Object}
\format{An \code{\link{R6Class}} generator object}
\usage{
testClass
}
\description{
Class for testing
}
\section{Methods}{

\itemize{
\item \strong{Constructor}
\itemize{
\item new: initialize a new \code{testClass} object
\itemize{
\item Arguments:
\itemize{
\item \code{N}: size of matrix
}
}
}
\item \strong{Methods}
\itemize{
\item testMethod: multiply the two matrix and return result
}
}
}

\keyword{R6}
\keyword{class}
