% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{testMethodCpp}
\alias{testMethodCpp}
\title{Test method for \code{\link{testClass}}}
\usage{
testMethodCpp(testClassEnv)
}
\description{
Multiplies two matrix together and returns result; should be equivalent to \code{testClass$testMethod()}
}
