% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{uniform}
\alias{uniform}
\title{Uniform distribution}
\usage{
uniform(min = 0, max = 1)
}
\arguments{
\item{min}{Lower bound of the distribution (default is zero)}

\item{max}{Upper bound of the distribution (default is one)}
}
\description{
Uniform distribution
}
